feat: added separate index page
